requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
aiosqlite
databases
pydantic
python-jose[cryptography]
passlib[bcrypt]
mcp-use
langchain-google-genai
python-dotenv
